Use signed integers for solution 8

+++ 8.c
@@ -10,8 +10,8 @@ int main() {
     int digits[DIGIT_COUNT];
     int digits_read = 0;
     int c = 0;
-    uint64_t largest_product = 0;
-    uint64_t current_product = 1;
+    int64_t largest_product = 0;
+    int64_t current_product = 1;
 
     while (digits_read < DIGIT_COUNT && (c = getchar()) != EOF) {
         if (isdigit(c)) {
@@ -26,8 +26,7 @@ int main() {
             current_product = 1;
 
             for (int d = 0; d < DIGIT_COUNT; d++) {
-                // Cast here is fine as a digit will always be unsigned
-                current_product *= (uint8_t) digits[d];
+                current_product *= digits[d];
             }
 
             // 2. Update largest product
@@ -48,7 +47,7 @@ int main() {
         } while (c != EOF);
     }
 
-    printf("%" PRIu64 "\n", largest_product);
+    printf("%" PRIi64 "\n", largest_product);
 
     return EXIT_SUCCESS;
 }
